Arreglos del merge(no esta todo arreglado)

+++ Anyaroth/Anyaroth/Poleaxe.cpp
@@ -17,11 +17,11 @@ void Poleaxe::update(const double & deltaTime)
 	{
 		_center = { _owner->getBody()->GetPosition().x * M_TO_PIXEL,_owner->getBody()->GetPosition().y * M_TO_PIXEL };
 		_pos = { _body->getBody()->GetPosition().x * M_TO_PIXEL, _body->getBody()->GetPosition().y * M_TO_PIXEL };
-		_nextPos = _pos + _velocity * _dir;
+		_nextPos = _pos + _velocity * (double)_dir;
 		if (_nextPos.distance(_center) > _offset.getX())
 		{
 			_dir = -_dir;
-			_nextPos = _pos + _velocity * _dir;
+			_nextPos = _pos + _velocity * (double)_dir;
 		}
 		_pos = _nextPos;
 		_body->getBody()->SetTransform(b2Vec2(_pos.getX() / M_TO_PIXEL, _pos.getY() / M_TO_PIXEL), 0);
